training/angela_lora/requirements.txt
# Angela LoRA Training - Requirements
# For MacBook Air M4 16GB with PyTorch MPS

# Core ML
torch>=2.1.0
transformers>=4.36.0
accelerate>=0.25.0
peft>=0.7.0
bitsandbytes>=0.41.0  # Note: limited MPS support, mainly for loading

# Data processing
datasets>=2.15.0
pandas>=2.0.0
asyncpg>=0.29.0  # For AngelaMemory database

# Tokenization
sentencepiece>=0.1.99
tiktoken>=0.5.0

# Utilities
tqdm>=4.66.0
wandb>=0.16.0  # Optional: for tracking
pyyaml>=6.0

# Evaluation
rouge-score>=0.1.2
nltk>=3.8.0
